requirements.txt
distro==1.5.0
numpy==1.18.4
pandas==1.0.3
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
tabula-py==2.1.0
